Adds initial files for temperature monitor
The base code for monitoring temperature in freezer has been
added. This code will log temperature and relative humidity data
every second to a log file. The log file rotates every day.

The code also drivers a 128x32 display to give read out of current
T/RH at the RPi. This display is set up to be off, and is turned on
for 30 seconds with a button press on GPIO 17.

To do:
  - Add code for generating plots
    * Generate at given time each day? Run in multiprocessing so as to
      not slow down main monitoring code

v0.0.0