Improve responsiveness of button, rolling data log
The button to turn on display is now instantaneous, with monitoring
of button handled in its own thread, which sents and event when
button pressed. A timer thread is also started to clear the
event so that display will turn off

The display is now in its own thread class as well, which updates
every one (1) second when the button event is set.

The 'main' thread polls the sensor at given interval, storing
data in rolling CSV file and updating the values in the display
class.

Class for rolling CSV file has been created. When date changes,
the current CSV file is closed, moved to new file name where
date is appended, then opens new CSV to begin writing again.
During the roll over, also look for files more than 30 days old
and delete them.

v0.1.0