"Fixes" web socket issue, better thread locks
There is an issue with the websocket client where if the server
is not found, the whole process hangs while it tries to connect.
This needs to be threaded better in the future using queues, but
no time now, so added the --no-socket flag for the CLI to disable
websocket code.

Have also added new decorator to the email code so that the
the threads spawned lock to ensure thread safe. Was hitting an
issue where email would send twice as first send was a bit slow,
so the variable holding time of last email send was not updated by the
time the second thread started. With the decorator, the first thing
that is done in the threads is to wait for the lock before sending
email, which ensures that timing is accurate.

Improves some logging and adds ability to set log level for log
file at CLI run.

v0.2.1