Adds number of polls check with finite
There was a issue with the first poll of a sensor returning NaN
and thus right after a boot-up would get email about all NaN.

Have added a module variable MIN_NUM_POLL = 10 and an 'nn' class
attribute to SHT30 that is incremented each time the poll() method
is called. We now check that self.nn > MIN_NUM_POLL before checking
average is finite.

Removes a time.sleep call in the display module on the button press
handler.